+++ src/main/java/lk/ijse/carrental/controller/CategoryFormController.java
@@ -7,15 +7,13 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Alert;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import lk.ijse.carrental.dto.CategoryDto;
+import lk.ijse.carrental.dto.CustomerDto;
 import lk.ijse.carrental.dto.tm.CategoryTm;
 import lk.ijse.carrental.service.custom.CategoryService;
 import lk.ijse.carrental.service.custom.ServiceFactory;
@@ -23,6 +21,7 @@ import lk.ijse.carrental.service.custom.ServiceType;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 
 public class CategoryFormController {
@@ -75,17 +74,23 @@ public class CategoryFormController {
         String id = txtCatId.getText();
         String name = txtCatName.getText();
 
-        var catDto = new CategoryDto(id, name);
+        if(!id.isEmpty() && !name.isEmpty()){
+            var catDto = new CategoryDto(id, name);
 
-        try {
-            categoryService.saveCategory(catDto);
-            new Alert(Alert.AlertType.CONFIRMATION,"Category Saved!").show();
-            clearFields();
-            initialize();
+            try {
+                categoryService.saveCategory(catDto);
+                new Alert(Alert.AlertType.CONFIRMATION,"Category Saved!").show();
+                clearFields();
+                initialize();
 
-        } catch (Exception e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+            } catch (Exception e) {
+                new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+            }
+        }else{
+            new Alert(Alert.AlertType.WARNING,"Please fill the required details!").show();
         }
+
+
 
     }
 
@@ -124,27 +129,41 @@ public class CategoryFormController {
         String id = txtCatId.getText();
         String name = txtCatName.getText();
 
-        var catDto = new CategoryDto(id,name);
-        try {
-            categoryService.updateCategory(catDto);
-            new Alert(Alert.AlertType.CONFIRMATION,"Category Updated!").show();
-            clearFields();
-            initialize();
-        } catch (Exception e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+        if(!id.isEmpty() && !name.isEmpty()){
+            var catDto = new CategoryDto(id,name);
+            try {
+                categoryService.updateCategory(catDto);
+                new Alert(Alert.AlertType.CONFIRMATION,"Category Updated!").show();
+                clearFields();
+                initialize();
+            } catch (Exception e) {
+                new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+            }
+        }else{
+            new Alert(Alert.AlertType.WARNING,"Please fill the required details!").show();
         }
     }
 
     public void btnDeleteOnAction(ActionEvent actionEvent) {
         String id = txtCatId.getText();
-        try {
-            CategoryDto catdto = categoryService.search(id);
-            categoryService.deleteCategory(catdto);
-            new Alert(Alert.AlertType.CONFIRMATION,"Category Deleted!").show();
-            clearFields();
-            initialize();
-        } catch (Exception e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+        if(!id.isEmpty()){
+            Alert alert = new Alert(Alert.AlertType.WARNING,
+                    "Do you want to delete this category?",
+                    ButtonType.YES, ButtonType.NO);
+            Optional<ButtonType> type = alert.showAndWait();
+            if(type.isPresent() && type.get() == ButtonType.YES) {
+                try {
+                    CategoryDto catdto = categoryService.search(id);
+                    categoryService.deleteCategory(catdto);
+                    new Alert(Alert.AlertType.CONFIRMATION, "Category Deleted!",ButtonType.OK).show();
+                    clearFields();
+                    initialize();
+                } catch (Exception e) {
+                    new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
+                }
+            }
+        }else {
+            new Alert(Alert.AlertType.WARNING,"Please select the required category id to delete!").show();
         }
     }
